+++ crates/ruff_python_semantic/src/lib.rs
@@ -4,6 +4,7 @@ mod branches;
 mod context;
 mod definition;
 mod globals;
+pub mod helpers;
 mod model;
 mod nodes;
 mod reference;
